fix(ci): resolve CI/CD workflow failures and React warnings
- Add conditional check for Docker Hub secrets to prevent login failure
- Fix Dockerfile ARG syntax for hadolint compatibility (remove trailing slash from GHCR_REGISTRY)
- Fix React Hook dependency warnings with eslint-disable comments for intentional patterns
- Extract hooks (useToast, useConfirm) and utilities (materialUtils, templateUtils) to separate files for Fast Refresh compatibility
