+++ frontend/src/components/shared/MaterialGeneratorModal.tsx
@@ -4,7 +4,7 @@ import { Image as ImageIcon, ImagePlus, Upload, X, FolderOpen } from 'lucide-rea
 import { Modal, Textarea, Button, useToast, MaterialSelector, Skeleton } from '@/components/shared';
 import { generateMaterialImage, getTaskStatus } from '@/api/endpoints';
 import { getImageUrl } from '@/api/client';
-import { materialUrlToFile } from './MaterialSelector';
+import { materialUrlToFile } from './materialUtils';
 import type { Material } from '@/api/endpoints';
 import type { Task } from '@/types';
