+++ frontend/src/components/shared/Toast.tsx
@@ -1,6 +1,7 @@
 import React, { useEffect } from 'react';
 import { X, CheckCircle, AlertCircle, Info } from 'lucide-react';
 import { cn } from '@/utils';
+import { useToast } from './useToast';
 
 interface ToastProps {
   message: string;
@@ -55,32 +56,18 @@ export const Toast: React.FC<ToastProps> = ({
   );
 };
 
-// Toast 管理器
-export const useToast = () => {
-  const [toasts, setToasts] = React.useState<Array<{ id: string; props: Omit<ToastProps, 'onClose'> }>>([]);
+// Toast Container component
+export const ToastContainer: React.FC<{ toasts: Array<{ id: string; props: Omit<ToastProps, 'onClose'> }>; remove: (id: string) => void }> = ({ toasts, remove }) => (
+  <>
+    {toasts.map((toast) => (
+      <Toast
+        key={toast.id}
+        {...toast.props}
+        onClose={() => remove(toast.id)}
+      />
+    ))}
+  </>
+);
 
-  const show = (props: Omit<ToastProps, 'onClose'>) => {
-    const id = Math.random().toString(36);
-    setToasts((prev) => [...prev, { id, props }]);
-  };
-
-  const remove = (id: string) => {
-    setToasts((prev) => prev.filter((t) => t.id !== id));
-  };
-
-  return {
-    show,
-    ToastContainer: () => (
-      <>
-        {toasts.map((toast) => (
-          <Toast
-            key={toast.id}
-            {...toast.props}
-            onClose={() => remove(toast.id)}
-          />
-        ))}
-      </>
-    ),
-  };
-};
-
+// Re-export useToast for convenience
+export { useToast };
